% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvac.R
\docType{data}
\name{rsvac}
\alias{rsvac}
\title{Repertoires of Sexual Violence in Armed Conflict (RSVAC) data}
\format{
\code{rsvac}
A data frame with 1,411 rows and 31 columns.
\describe{
\item{year}{Year.}
\item{conflictid_old}{Old UCDP/PRIO Conflict ID (before version 17.1). From UCDP/PRIO.}
\item{conflictid_new}{New UCDP/PRIO Conflict ID (starting with version 17.1 of UCDP data). From UCDP/PRIO.}
\item{actor}{Name of country if the actor is a government, otherwise name of organization if this is a rebel organization or militia. From UCDP/PRIO.}
\item{actorid}{Old UCDP Non-State Actor ID (before version 17.1). From UCDP/PRIO.}
\item{actorid_new}{In version 17.1 of all UCDP data ID the system for conflicts, actors and dyads was changed in order to make them unique across all UCDP core data sets and all UCDP types of violence. From UCDP/PRIO.}
\item{actor_type}{Type of actor: 1 - state or incumbent government ("Side A" in UCDP dyadic), 2 - State A2 ("Side A2nd" in UCDP dyadic), states supporting State 1 involved with the conflict on its territory, 3 - Rebel ("Side B" in UCDP dyadic), 4 - State supporting "Side B" in other country ("Side B2nd" in UCDP dyadic), 5 - Second state in interstate conflict ("Side B" in UCDP dyadic), 6 - Pro-government militias. From SVAC.}
\item{type}{Conflict type: 2 - interstate conflict, 3 - intrastate conflict, 4 - internationalized armed conflict. From UCDP/PRIO.}
\item{incomp}{Conflict issue: 1 - territory, 2 - government, 3 - government and territory. From UCDP/PRIO.}
\item{region}{Numeric coding of geographic region. From UCDP/PRIO.}
\item{location}{The name(s) of the country/countries of fighting and whose government(s) have a primary claim to the territory in dispute. From UCDP/PRIO.}
\item{gwnoloc}{Gleditsch/Ward country ID of location variable 1.}
\item{gwnoloc2}{Gleditsch/Ward country ID of location variable 2.}
\item{gwnoloc3}{Gleditsch/Ward country ID of location variable 3.}
\item{gwnoloc4}{Gleditsch/Ward country ID of location variable 4.}
\item{conflictyear}{Binary indicator of an active conflict-year. From UCDP/PRIO.}
\item{interm}{Binary indicator of an post-conflict-year. From UCDP/PRIO.}
\item{postc}{Binary indicator of an active conflict-year. From UCDP/PRIO.}
\item{svac_form}{Reported forms of sexual violence coded in the SVAC data set. Variable shows which forms were coded using the SVAC definitions of sexual violence. This variable is included to preserve transparency and comparability of SVAC and RSVAC. Users of the RSVAC dataset should rely on \code{rsvac_form} to determine the forms an actor engaged in.}
\item{rsvac_form}{Reported forms of sexual violence coded in the RSVAC data set. Variable was coded using definitions in Section 2.0 of the RSVAC Coding Manual. "-99" is used where SVAC found reports of sexual violence but RSVAC found no reports of any form.}
\item{form}{The form of sexual violence being described. One of: forced abortion and forced sterilization, forced pregnancy (no instances in the data set), forced prostitution, rape, sexual abuse, sexual mutilation, sexual slavery and forced marriage, or sexual torture.}
\item{state_prev}{Ordinal prevalence measure from 0--3 using the annual U.S. State Department "Country Report on Human Rights Practices". \code{NA} indicates no report found and no data available.}
\item{ai_prev}{Ordinal prevalence measure from 0--3 using the annual reports and special topical reports of Amnesty International. \code{NA} indicates no report found and no data available.}
\item{hrw_prev}{Ordinal prevalence measure from 0--3 using the annual reports and special topical reports of Human Rights Watch. \code{NA} indicates no report found and no data available.}
\item{mp}{Ordinal prevalence measure of multiple perpetrator incidents from 0--2 using U.S. State Department, Amnesty International, and Human Rights Watch reports.}
\item{mp_notes}{A qualitative note that summarizes the reporting on multiple perpetrator incidents.}
\item{victim_notes}{A qualitative note that summarizes reported information on victim characteristics when information is available. Additionally, this field includes notation to disaggregate reports of sexual slavery and forced marriage and reports of forced abortion and forced sterilization.}
\item{selection_notes}{A qualitative note that summarizes reported information on victim targeting when information is available.}
\item{timing_notes}{A qualitative note that summarizes reported information on the timing of sexual violence when information is available.}
\item{location_notes}{A qualitative note that summarizes reported information on the location of sexual violence when information is available.}
\item{witness_notes}{A qualitative note that summarizes reported information on witnesses of sexual violence when information is available}
}
}
\source{
\doi{doi:10.1177/00223433211044674}
}
\usage{
rsvac
}
\description{
Repertoires of Sexual Violence in Armed Conflict (RSVAC) data set
(Dumaine et al. 2022) \doi{doi:10.1177/00223433211044674}. Data covers the period 1989--2015.
}
\keyword{datasets}
